half-marathon-bruin/pipeline/assets/ingestion/requirements.txt
kagglehub[pandas-datasets]>=1.0.0
google-cloud-storage>=2.16.0
google-cloud-bigquery>=3.17.0
db-dtypes>=1.2.0
pandas>=2.0.0